[LayoutNG] Invalidate layout when OOF-positioned gets added/removed.
This patch allows us to reuse a cached layout result when we have
OOF-positioned descendants in the result.

In order to do this whenever a OOF-positioned object gets added or
removed we invalidate the parent chain for all objects which may
contain this element inside its layout result.

There are more optimal versions of this approach, which I've noted
inside a comment, but this seems good enough for now.

Bug: 635619
Change-Id: I4b29fcc26154dcd76cee55455d83558c6cf798fa