WPT: service worker: Loosen postmessage-to-client-message-queue.https.html.
Chrome flakily failed the test because it doesn't guarantee the order
between message and fetch event. That might violate the spec, but the
test doesn't need to test this. It's meant to test that the client
message queue isn't enabled until a certain time. We can test
that without requiring message/fetch ordering.

Loosen the test to just check that the messages arrive after
|earliest_dispatch|.

Bug: 926670
Change-Id: Ic291b25443358412936b4a5eeb24ee6f8efa6b71